requirements.txt
astropy==2.0.1
matplotlib==2.0.2
progressbar2==3.34.2
pyquery==1.2.17
requests==2.18.3
